+++ src/sabledocs/sable_config.py
@@ -24,7 +24,7 @@ class SableConfig:
         self.repository_type = RepositoryType.NONE
         self.ignore_comments_after: List[str] = []
         self.ignore_comment_lines_containing: List[str] = []
-        self.hide_packages: List[str] = []
+        self.hidden_packages: List[str] = []
 
         if path.exists(config_file_path):
             print(f"Configuration found in {config_file_path}")
@@ -60,6 +60,6 @@ class SableConfig:
 
                 self.ignore_comment_lines_containing = config_values.get('ignore-comment-lines-containing', [])
 
-                self.hide_packages = config_values.get('hide-packages', [])
+                self.hidden_packages = config_values.get('hidden-packages', [])
         else:
             print("sabledocs.toml file not found, using default configuration.")