+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unpublished
 - Packages can now be hidden. Added
-  - option `hide-packages` to the config and
+  - option `hidden-packages` to the config and
   - iterable `non_hidden_packages` to be used in the templates `package.html` and `index.html`
 
 ## [0.6]